add list.other
